A form of DDA now exists. Aims for 50%, but due to low number of evaluations a margin of 10% is in place. Current goal is achieving 50% +- 10% within 5 evaluations and is working properly with mouse testing. TODO: test more in VR, it's very different from testing with a mouse

+++ MoneyMoves/Assets/Scripts/Minigames/Whac-A-Sphere/WhacASphereVariables.cs
@@ -8,6 +8,6 @@ public class WhacASphereVariables : ScriptableObject
     public float activeTime;
     public float negativeActiveTime;
     public float timeBetweenActivation;
-    public int totalTime;
+    public float totalTime;
     public int spawnNegativeAfterSpawns;
 }
